pm: device_runtime: detect power up failures
Check that the device we are trying to use has successfully powered up
when it is on a power domain. If it has not, release the request for the
power domain to be powered to ensure we do not have dangling requests.

As a result of this, `PM_DEVICE_ACTION_RESUME` will not be run on a
device if `PM_DEVICE_ACTION_TURN_ON` has already failed.

(cherry picked from commit 962eca65aff2c03e7063d50b541b8f735b1a34cf)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: 962eca65aff2c03e7063d50b541b8f735b1a34cf
Change-Id: Ie6b23d069c9f9b37b08e8cf6e649d81ab7498eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3757569
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>