scripts: add github_helpers.py
This is meant to be a place where we can store generic zephyr wrappers
around the de-facto standard github API for python.

The first 'customer' will be a script that snapshots our open bugs at a
particular point in time, which will be added in a later patch.

I think it's useful to factor this file out of there from the
beginning just to keep things clean, even though I don't have a second
customer in mind at the moment.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 8aaea6d200c190f7f28566616aaaf3b56c182e75
Change-Id: I3066d03c343ddbd4c563b9c5ed37b37bd5652abc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3667445
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>