doc: Add LwM2M engine state machine
The state machine depicts the states of the lwm2m engine
and the spawned RD client events that will be forwarded
to lwm2m client if it has registered a callback-function.
Most of the events are just informs for the client and the
RD client underneath works and recovers automatically from
most cases.

Signed-off-by: Jarno Lamsa <jarno.lamsa@nordicsemi.no>
GitOrigin-RevId: e0e9ec7714e4c43290958f058c34e4ba133fd823
Change-Id: I61bb9ef90bd548cf12d92779e152a2ab8d887faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3657371
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>