twister: rework how we discard tests
Add a filter type to distinguish between testsuite filters that are
maintained in the testcase.yaml file and those provided on the command
line.

The issue was that when something is excluded deliberately on the command
line twister would report an error if that filter would filter out an
integration platform. We do not want that, because the filtering is
being done by the caller knowing that some tests would be not run, even
if those are integration platforms.

(cherry picked from commit 87be2f591f293388ae18b9a4cf962d03f9593580)

Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 87be2f591f293388ae18b9a4cf962d03f9593580
Change-Id: I70d19793025c022649ec11f12f2008082777c958
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3699522
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>