boards: arm: arduino_nicla_sense_me: Add user button
This commit adds a user button to the board DTS.

The button is not connected to nRESET as suggested
by the board schematics, but rather connected to
P0.21. The source code found in the Arduino IDE
connects an interrupt to this GPIO and calls
NVIC_System_Reset from there to cause the reset.

The button was therefore added as a regular button
to cover the most general case.

This has been verified on an Arduino Nicla Sense ME.

(cherry picked from commit 48e905c7316ca64769578b221811e6fc02cf2441)

Original-Signed-off-by: Benjamin Björnsson <benjamin.bjornsson@gmail.com>
GitOrigin-RevId: 48e905c7316ca64769578b221811e6fc02cf2441
Change-Id: Ia5de12667685d5464834844202d7a3deb8b1b927
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3720850
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>