tests/kernel/fatal: Work around historical API misuse
This test was written to do a k_oops() in the main thread.  That's an
essential thread, and aborting it is actually a system panic now.  The
test was written contra the docs on this, but it worked fine for
years.

Just reflag the thread as a simple workaround rather than trying
anything fancy. This is a very simple test.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: c64ac967a85f2eb5acaf3ac901979594565fc2fe
Change-Id: I2fb9af27812c03d10c95e6c00ac734ae9e3fa9f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3657585
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>