dts: arm: stm32h7 devices have spi clock source on pll1_q
All the SPI1,2,3 clock of the stm32h7x mcus
are sourced by the PLL1_Q by default. This must be set
in the DTS to have a valid clock rate calculation.
The pll1_q is divided by the <div-q> property of the pll node.

Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: fd6a3d3154c96f08c75143373d31ea5eb47c1756
Change-Id: I3e962d75956e4dc65c4c5418c278873dbb0bd3d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3667421
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>