civetweb: remove obsolete code
This code has gone unmaintained and bugs continue to be reported
against it. We do not have the resources as a project to maintain this
in "odd fixes" mode, and nobody has stepped up to maintain it [1], so
sadly this must be removed for now.

If anyone would like to see civetweb supported in upstream Zephyr
again, they are welcome to add it back, as long as they promise to
maintain it going forward.

Many thanks to everyone who has contributed to civetweb support in
Zephyr while it was here. So long and thanks for all the fish.

[1] https://lists.zephyrproject.org/g/devel/message/8466

(cherry picked from commit f49d11d310367dd8b6eb04c3fa73c7d8c1789ad4)

Original-Fixes: #45807
Original-Fixes: #43910
Original-Fixes: #34226
Original-Fixes: #46743
Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: f49d11d310367dd8b6eb04c3fa73c7d8c1789ad4
Change-Id: I441c6907595335435d31a713d3c162571b40a741
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3720852
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>