libc: minimal: Add PRIxMAX macros for [u]intmax_t
This commit adds the missing `PRIxMAX` macros for the C99 `intmax_t`
and `uintmax_t` types:

  PRIdMAX, PRIiMAX, PRIoMAX, PRIuMAX, PRIxMAX, PRIXMAX

Note that the `PRIxMAX` macros specify the `ll` size modifier because
the type of the `intmax_t` for the minimal libc is defined as that of
the `int64_t`, which is always overridden to `long long int` by
`zephyr_stdint.h`; for more details, refer to the GitHub PR #29876,
which deliberately introduced this scheme.

In the future, this scheme will need to be reworked such that the
minimal libc `stdint.h` defines `intmax_t` as `__INTMAX_TYPE__`, and
the `inttypes.h` resolves the corresponding format specifier.

(cherry picked from commit d755050a14c4648d1cfc47e912d81d153381490c)

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: d755050a14c4648d1cfc47e912d81d153381490c
Change-Id: I851871e954940c7c0aa34be2efcdf9acb39adcfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3672267
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>