cmake: modules: deprecate civetweb
This module has gone unmaintained and we do not have the resources to
respond to bug reports for it.

The TSC has decided to deprecate it for Zephyr v3.1 and remove it
entirely in Zephyr v3.2 unless a maintainer volunteers to keep this
module up to date and fix bugs.

This commit generates a warning at CMake time if this module is
compiled in. The warning will appear in the cmake output in the same
place that deprecated board or SoC warnings would.

(cherry picked from commit 013a0034c90a3b5d5b459263ee89bc5ed7582596)

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 013a0034c90a3b5d5b459263ee89bc5ed7582596
Change-Id: I95eebf524ba3d16e19d7b48ba99e92095af5407b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3675227
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>