mgmt/osdp: phy: Rework MARK byte handling
OSDP specification section 5.7 states that a transmitting device has to
drive the transmission line to a marking state for a period of one char
in the current baud rate. This can be achieved by sending 0xFF. Since
this is not mentioned in the packet structure definition, many commercial
implementations of OSDP out in the wild do not send/expect this byte.

To work with such non-conforming devices, we will try to be as flexible
as possible in the PD: send mark byte only if the other side sent one. In
case of CP, we have no option but to send the mark byte to be as close
to the specification as possible. If a particular use case needs the CP
to not send it, we will provide a Kconfig option to disable it.

(cherry picked from commit a233dea2852b459e7a0b0ad76d7c04d9727a1ef4)

Original-Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>
GitOrigin-RevId: a233dea2852b459e7a0b0ad76d7c04d9727a1ef4
Change-Id: I7461988e393c9a3a9cf4a94b315537f1aa61c981
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4200356
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>